pyproject.toml
[project]
name = "sgvmon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
]
